moving the animation to utils
